fix: move errors to serializer
